+++ src/components/Movies/index.js
@@ -1,0 +1,1 @@
+export { default } from "./Movies";
