+++ src/layouts/AppLayout.jsx
@@ -1,6 +1,6 @@
 import React from "react";
-import Header from "../components/Header";
 import { Outlet } from "react-router-dom";
+import Header from "../components/Header";
 
 const AppLayout = () => {
   return (
